+++ base/deprecated.jl
@@ -90,7 +90,8 @@ function firstcaller(bt::Array{Ptr{Void},1}, funcsyms)
     lkup = StackTraces.UNKNOWN
     for frame in bt
         lkups = StackTraces.lookup(frame)
-        for lkup in lkups
+        for l in lkups
+            lkup = l
             if lkup == StackTraces.UNKNOWN
                 continue
             end
@@ -1067,7 +1068,8 @@ function partial_linear_indexing_warning_lookup(nidxs_remaining)
         caller = StackTraces.UNKNOWN
         for frame in bt
             lkups = StackTraces.lookup(frame)
-            for caller in lkups
+            for c_ in lkups
+                caller = c_
                 if caller == StackTraces.UNKNOWN
                     continue
                 end